requirement.txt
allure-behave==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0
behave==1.3.3
certifi==2025.10.5
colorama==0.4.6
cucumber-expressions==18.0.1
cucumber-tag-expressions==8.0.0
Faker==37.12.0
h11==0.16.0
idna==3.11
outcome==1.3.0.post0
parse==1.20.2
parse_type==0.6.6
pluggy==1.6.0
PySocks==1.7.1
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.2.0
